Add PostMapping producto

+++ src/main/java/com/morealva/service/impl/ProductoServiceImpl.java
@@ -117,13 +117,14 @@ public class ProductoServiceImpl extends CRUDImpl<Producto, Integer> implements 
         return producto;
     }
 
+    @Transactional
     @Override
-    public Producto saveTransactional(Producto producto, List<PrincipioActivo> principio){
-        if (producto.getIdProducto() != null) {
+    public Producto saveTransactional(Producto producto, List<PrincipioActivo> principios){
+        if (producto.getIdProducto() != null && producto.getIdProducto()>0) {
             productoPrincipioActivoRepo.deletePrincipioActivo(producto.getIdProducto());
         }
         productoRepo.save(producto);
-        principio.forEach(ex -> productoPrincipioActivoRepo.savePrincipioActivo(producto.getIdProducto(), ex.getIdPrincipio()));
+        principios.forEach(ex -> productoPrincipioActivoRepo.savePrincipioActivo(producto.getIdProducto(), ex.getIdPrincipio()));
         return producto;
     }
 }
